requirements.txt
dj-database-url==0.3.0
Django==1.8.1
gunicorn==19.3.0
SQLAlchemy==1.0.4
whitenoise==1.0.6
python-telegram-bot==2.8.3
